+++ src/Builder/Resources/Contact/Batch.php
@@ -6,49 +6,61 @@
  * @link       https://github.com/flipbox/relay-hubspot
  */
 
-namespace Flipbox\Relay\HubSpot\Builder\Resources\Limit\Daily;
+namespace Flipbox\Relay\HubSpot\Builder\Resources\Contact;
 
 use Flipbox\Relay\HubSpot\AuthorizationInterface;
 use Flipbox\Relay\HubSpot\Builder\HttpRelayBuilder;
+use Flipbox\Relay\HubSpot\Middleware\JsonRequest as JsonMiddleware;
 use Flipbox\Relay\HubSpot\Middleware\ResourceV1;
-use Flipbox\Relay\Middleware\SimpleCache as CacheMiddleware;
 use Psr\Log\LoggerInterface;
-use Psr\SimpleCache\CacheInterface;
 
 /**
  * @author Flipbox Factory <hello@flipboxfactory.com>
- * @since 1.0.2
+ * @since 1.0.3
  */
-class Read extends HttpRelayBuilder
+class Batch extends HttpRelayBuilder
 {
     /**
      * The node
      */
-    const NODE = 'integrations';
+    const NODE = 'contacts';
 
     /**
      * The resource
      */
-    const RESOURCE = 'limit/daily';
+    const RESOURCE = 'contact/batch';
 
     /**
+     * Upsert constructor.
+     * @param array $payload
      * @param AuthorizationInterface $authorization
-     * @param CacheInterface $cache
      * @param LoggerInterface|null $logger
      * @param array $config
      */
     public function __construct(
+        array $payload,
         AuthorizationInterface $authorization,
-        CacheInterface $cache,
         LoggerInterface $logger = null,
         $config = []
     ) {
         parent::__construct($authorization, $logger, $config);
 
-        $cacheKey = self::RESOURCE;
-
         $this->addUri($logger)
-            ->addCache($cache, $cacheKey, $logger);
+            ->addPayload($payload, $logger);
+    }
+
+    /**
+     * @param array $payload
+     * @param LoggerInterface|null $logger
+     * @return $this
+     */
+    protected function addPayload(array $payload, LoggerInterface $logger = null)
+    {
+        return $this->addAfter('body', [
+            'class' => JsonMiddleware::class,
+            'payload' => $payload,
+            'logger' => $logger ?: $this->getLogger()
+        ], 'uri');
     }
 
     /**
@@ -59,25 +71,10 @@ class Read extends HttpRelayBuilder
     {
         return $this->addBefore('uri', [
             'class' => ResourceV1::class,
+            'method' => 'POST',
             'node' => self::NODE,
             'resource' => self::RESOURCE,
             'logger' => $logger ?: $this->getLogger()
         ]);
-    }
-
-    /**
-     * @param CacheInterface $cache
-     * @param string $key
-     * @param LoggerInterface|null $logger
-     * @return $this
-     */
-    protected function addCache(CacheInterface $cache, string $key, LoggerInterface $logger = null)
-    {
-        return $this->addBefore('cache', [
-            'class' => CacheMiddleware::class,
-            'logger' => $logger ?: $this->getLogger(),
-            'cache' => $cache,
-            'key' => $key
-        ], 'token');
     }
 }